Simplify prompt shortcut handling
