+++ packages/app/src/keyboard/keyboard-shortcuts.ts
@@ -17,7 +17,6 @@ export type KeyboardShortcutContext = {
   focusScope: KeyboardFocusScope;
   commandCenterOpen: boolean;
   hasSelectedAgent: boolean;
-  agentAwaitingInput: boolean;
 };
 
 export type KeyboardShortcutMatch = {
@@ -66,8 +65,6 @@ interface ShortcutWhen {
   commandCenter?: false;
   /** true = requires a selected agent */
   hasSelectedAgent?: true;
-  /** true = requires a visible prompt awaiting agent input */
-  agentAwaitingInput?: true;
   /** Exact focus scope match */
   focusScope?: KeyboardFocusScope;
 }
@@ -822,7 +819,7 @@ const SHORTCUT_BINDINGS: readonly ShortcutBinding[] = [
     action: "agent.prompt.select",
     combo: "1",
     repeat: false,
-    when: { agentAwaitingInput: true, focusScope: "other", commandCenter: false },
+    when: { focusScope: "other", commandCenter: false },
     payload: { type: "index" },
     help: {
       id: "agent-prompt-select-1",
@@ -836,7 +833,7 @@ const SHORTCUT_BINDINGS: readonly ShortcutBinding[] = [
     action: "agent.prompt.select",
     combo: "2",
     repeat: false,
-    when: { agentAwaitingInput: true, focusScope: "other", commandCenter: false },
+    when: { focusScope: "other", commandCenter: false },
     payload: { type: "index" },
     help: {
       id: "agent-prompt-select-2",
@@ -850,7 +847,7 @@ const SHORTCUT_BINDINGS: readonly ShortcutBinding[] = [
     action: "agent.prompt.select",
     combo: "3",
     repeat: false,
-    when: { agentAwaitingInput: true, focusScope: "other", commandCenter: false },
+    when: { focusScope: "other", commandCenter: false },
     payload: { type: "index" },
     help: {
       id: "agent-prompt-select-3",
@@ -1009,7 +1006,6 @@ function matchesWhen(when: ShortcutWhen | undefined, context: KeyboardShortcutCo
   if (when.terminal === false && context.focusScope === "terminal") return false;
   if (when.commandCenter === false && context.commandCenterOpen) return false;
   if (when.hasSelectedAgent === true && !context.hasSelectedAgent) return false;
-  if (when.agentAwaitingInput === true && !context.agentAwaitingInput) return false;
   if (when.focusScope !== undefined && context.focusScope !== when.focusScope) return false;
   return true;
 }
