+++ packages/app/src/panels/agent-panel.tsx
@@ -22,11 +22,9 @@ import { useArchiveAgent } from "@/hooks/use-archive-agent";
 import { useDelayedHistoryRefreshToast } from "@/hooks/use-delayed-history-refresh-toast";
 import { useAgentInputDraft } from "@/hooks/use-agent-input-draft";
 import { useKeyboardShiftStyle } from "@/hooks/use-keyboard-shift-style";
-import { useKeyboardActionHandler } from "@/hooks/use-keyboard-action-handler";
 import { useStableEvent } from "@/hooks/use-stable-event";
 import { usePaneContext } from "@/panels/pane-context";
 import type { PanelDescriptor, PanelRegistration } from "@/panels/panel-registry";
-import type { KeyboardActionDefinition } from "@/keyboard/keyboard-action-dispatcher";
 import {
   useHostRuntimeClient,
   useHostRuntimeConnectionStatus,
@@ -323,25 +321,6 @@ function AgentPanelBody({
     requiresAttention: agent?.requiresAttention,
     attentionReason: agent?.attentionReason,
     isScreenFocused: isPaneFocused,
-  });
-
-  const handlePromptShortcut = useCallback(
-    (action: KeyboardActionDefinition): boolean => {
-      if (!isPaneFocused || action.id !== "agent.prompt.select") {
-        return false;
-      }
-      return streamViewRef.current?.selectPendingPermissionOption(action.index) ?? false;
-    },
-    [isPaneFocused],
-  );
-
-  useKeyboardActionHandler({
-    handlerId: `agent-prompt-actions:${serverId}:${agentId ?? "__pending__"}`,
-    actions: ["agent.prompt.select"],
-    enabled: Boolean(agentId),
-    priority: 150,
-    isActive: () => isPaneFocused,
-    handle: handlePromptShortcut,
   });
 
   useEffect(() => {
@@ -818,6 +797,7 @@ function AgentPanelBody({
                 agent={effectiveAgent}
                 streamItems={shouldUseOptimisticStream ? mergedStreamItems : streamItems}
                 pendingPermissions={pendingPermissions}
+                isPaneFocused={isPaneFocused}
                 routeBottomAnchorRequest={routeBottomAnchorRequest}
                 isAuthoritativeHistoryReady={hasAppliedAuthoritativeHistory}
                 onOpenWorkspaceFile={onOpenWorkspaceFile}
